moved text values formatting from DMN model to converters
