+++ modules/flowable-dmn-model/src/main/java/org/flowable/dmn/model/UnaryTests.java
@@ -39,8 +39,6 @@ public class UnaryTests extends DmnElement {
 
     public void setTextValues(List<Object> textValues) {
         this.textValues = textValues;
-        this.text = "\"" + StringUtils.join(textValues, "\",\"") + "\"";
-
     }
 
     public String getExpressionLanguage() {
